src/requirements/base.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0

# db
sqlmodel==0.0.8
asyncpg==0.27.0
alembic==1.8.1
pydantic[email]==1.10.2
sqlalchemy-utils==0.38.3
fastapi-async-sqlalchemy==0.3.12
aioredis==2.0.1
fastapi-cache2[redis]==0.1.9

# utils
loguru==0.6.0
environs==9.5.0
orjson==3.8.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5

aiohttp==3.8.3
